requirements.txt
Flask==2.1.3
flask-cors==3.0.10
Flask-PyMongo==2.3.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
PyJWT==2.1.0
python-dotenv==0.19.2
